refactor: stake and unstake modal outside of BalancesContext
