+++ src/app/user/Balances/StakeRIFCell.tsx
@@ -1,14 +1,16 @@
 import { useBalancesContext } from '@/app/user/Balances/context/BalancesContext'
 import { Button } from '@/components/Button'
+import { useModal } from '@/shared/hooks/useModal'
 import { useSearchParams } from 'next/navigation'
 import { useEffect, useState } from 'react'
 
 export const StakeRIFCell = () => {
+  const stakeModal = useModal()
   const searchParams = useSearchParams()
   const action = searchParams?.get('action')
   const [openModal, setOpenModal] = useState(action === 'stake')
 
-  const { stakeModal, balances } = useBalancesContext()
+  const { balances } = useBalancesContext()
   const { balance } = balances['RIF']
   const hasEnoughBalance = Number(balance) > 0
