+++ src/app/user/Balances/UnStakeRIFCell.tsx
@@ -1,8 +1,10 @@
 import { useBalancesContext } from '@/app/user/Balances/context/BalancesContext'
 import { Button } from '@/components/Button'
+import { useModal } from '@/shared/hooks/useModal'
 
 export const UnStakeRIFCell = () => {
-  const { unstakeModal, balances } = useBalancesContext()
+  const unstakeModal = useModal()
+  const { balances } = useBalancesContext()
   const { balance } = balances['stRIF']
   const hasEnoughBalance = Number(balance) > 0
   return (
