+++ src/app/user/Stake/StakingSteps.tsx
@@ -7,6 +7,7 @@ import { StakingToken } from '@/app/user/Stake/types'
 import { tokenContracts } from '@/lib/contracts'
 import { stakingSteps } from './Steps/stepsUtils'
 import { useStakeRIF } from '@/app/user/Stake/hooks/useStakeRIF'
+import { useModal } from '@/shared/hooks/useModal'
 
 interface StakingStepsProps {
   onCloseModal: () => void
@@ -56,7 +57,7 @@ const StakingSteps = ({ onCloseModal }: StakingStepsProps) => {
 }
 
 export const StakingModal = () => {
-  const { stakeModal } = useBalancesContext()
+  const stakeModal = useModal()
 
   if (stakeModal.isModalOpened) {
     return <StakingSteps onCloseModal={stakeModal.closeModal} />
