+++ src/app/user/Stake/UnStakingSteps.tsx
@@ -7,6 +7,7 @@ import { tokenContracts } from '@/lib/contracts'
 import { StakingProvider } from '@/app/user/Stake/StakingContext'
 import { Modal } from '@/components/Modal/Modal'
 import { useUnstakeStRIF } from '@/app/user/Stake/hooks/useUnstakeStRIF'
+import { useModal } from '@/shared/hooks/useModal'
 
 interface StakingStepsProps {
   onCloseModal: () => void
@@ -57,7 +58,7 @@ const UnStakingSteps = ({ onCloseModal }: StakingStepsProps) => {
 }
 
 export const UnStakingModal = () => {
-  const { unstakeModal } = useBalancesContext()
+  const unstakeModal = useModal()
 
   if (unstakeModal.isModalOpened) {
     return <UnStakingSteps onCloseModal={unstakeModal.closeModal} />
